move LLVMOptimizer from tool to llvm_util
