+++ tools/alive-tv.cpp
@@ -2,9 +2,9 @@
 // Distributed under the MIT license that can be found in the LICENSE file.
 
 #include "llvm_util/llvm2alive.h"
+#include "llvm_util/LLVMOptimizer.h"
 #include "smt/smt.h"
 #include "tools/transform.h"
-#include "tools/LLVMOptimizer.h"
 #include "util/version.h"
 
 #include "llvm/ADT/StringExtras.h"
